Handle mutations in GraphiQL
Remove unused files
