+++ .gitignore
@@ -4,3 +4,4 @@ build
 !./dist/.gitkeep
 *.zip
 .DS_Store
+development/server/db.json
