+++ development/server/db.json
@@ -1,9 +1,3 @@
 {
-  "favorites": [
-    {
-      "name": "Spectra",
-      "hex": "#32605D",
-      "contrast": "#ffffff"
-    }
-  ]
+  "favorites": []
 }
